Класс ExampleUser разделен на 2 сущности, добавлена типизация
