+++ README.md
@@ -116,6 +116,7 @@ $User->explainAccess('/ваш/маршрут');
 * Сделать объекты конфигурабельными.
 * Проверить возможность установки явного времени действия сессии (авторизация).
 * Сделать обработку ситуации отсутствия доступа более "мягкой" (без вообще непрехватываемого исключения).
+* Добавить возможность редактирования роли пользователя, проверить стили на странице редактирования.
