+++ application/config/web.php
@@ -31,7 +31,7 @@ $config = [
 		=> \application\handlers\UserExceptionHandler::class
         ],
         'user' => [ // подсистема авторизации
-            'class' => \application\models\ExampleUser::class,
+            'class' => \application\models\AuthUser::class,
 	    'construct' => [
                 'session' => '@session',
                 'router' => '@router'
