+++ application/models/UserModel.php
@@ -1,0 +1,130 @@
+<?php
+namespace application\models;
+
+use ItForFree\SimpleMVC\mvc\Model;
+/**
+ * Класс для обработки пользователей
+ */
+class UserModel extends Model
+{
+    // Свойства
+    /**
+    * @var string логин пользователя
+    */
+    public $login = null;
+    
+    public ?int $id = null;
+
+    /**
+    * @var string пароль пользователя
+    */
+    public $pass = null;
+    
+    /**
+    * @var string роль пользователя
+    */
+    protected $role = null;
+    
+    public $email = null;
+    
+    public $timestamp = null;
+    
+    /**
+     * @var string Критерий сортировки строк таблицы
+     */
+    public string $orderBy = "login ASC";
+    
+    /**
+     *  @var string название таблицы
+     */
+    public string $tableName = 'users';
+    
+    public $salt = null;
+    
+
+    public function insert()
+    {
+        $sql = "INSERT INTO $this->tableName (timestamp, login, salt, pass, role, email) VALUES (:timestamp, :login, :salt, :pass, :role, :email)"; 
+        $st = $this->pdo->prepare ( $sql );
+        $st->bindValue( ":timestamp", (new \DateTime('NOW'))->format('Y-m-d H:i:s'), \PDO::PARAM_STMT);
+        $st->bindValue( ":login", $this->login, \PDO::PARAM_STR );
+        
+        //Хеширование пароля
+        $this->salt = rand(0,1000000);
+        $st->bindValue( ":salt", $this->salt, \PDO::PARAM_STR );
+//        \DebugPrinter::debug($this->salt);
+        
+        $this->pass .= $this->salt;
+        $hashPass = password_hash($this->pass, PASSWORD_BCRYPT);
+//        \DebugPrinter::debug($hashPass);
+        $st->bindValue( ":pass", $hashPass, \PDO::PARAM_STR );
+        
+        $st->bindValue( ":role", $this->role, \PDO::PARAM_STR );
+        $st->bindValue( ":email", $this->email, \PDO::PARAM_STR );
+        $st->execute();
+        $this->id = $this->pdo->lastInsertId();
+    }
+    
+    public function update()
+    {
+        $sql = "UPDATE $this->tableName SET timestamp=:timestamp, login=:login, pass=:pass, email=:email  WHERE id = :id";  
+        $st = $this->pdo->prepare ( $sql );
+        
+        $st->bindValue( ":timestamp", (new \DateTime('NOW'))->format('Y-m-d H:i:s'), \PDO::PARAM_STMT);
+        $st->bindValue( ":login", $this->login, \PDO::PARAM_STR );
+        
+        // Хеширование пароля
+        $this->salt = rand(0,1000000);
+        //$st->bindValue( ":salt", $this->salt, \PDO::PARAM_STR );
+        //$this->pass .= $this->salt;
+        //$hashPass = password_hash($this->pass, PASSWORD_BCRYPT);
+        $st->bindValue( ":pass", $this->pass, \PDO::PARAM_STR );
+        
+        //$st->bindValue( ":role", $this->role, \PDO::PARAM_STR );
+        $st->bindValue( ":email", $this->email, \PDO::PARAM_STR );
+        $st->bindValue( ":id", $this->id, \PDO::PARAM_INT );
+        $st->execute();
+    }
+    
+    /**
+     * Вернёт id пользователя
+     * 
+     * @return ?int
+     */
+    public function getId()
+    {
+        if ($this->userName !== 'guest'){
+            $sql = "SELECT id FROM users where login = :userName";
+            $st = $this->pdo->prepare($sql); 
+            $st -> bindValue( ":userName", $this->userName, \PDO::PARAM_STR );
+            $st -> execute();
+            $row = $st->fetch();
+            return $row['id']; 
+        } else  {
+            return null;
+        }  
+    }
+    
+    /**
+     * Проверка логина и пароля пользователя.
+     */
+    public function checkAuth($login) {
+	$sql = "SELECT salt, pass FROM users WHERE login = :login";
+	$st = $this->pdo->prepare($sql);
+	$st->bindValue(":login", $login, \PDO::PARAM_STR);
+	$st->execute();
+	return $st->fetch();
+    }
+    
+    /**
+     * Проверяем активность пользователя.
+     */
+    public function getRole($login) {
+	$sql = "SELECT role FROM users WHERE login = :login";
+	$st = $this->pdo->prepare($sql);
+	$st->bindValue(":login", $login, \PDO::PARAM_STR);
+	$st->execute();
+	return $st->fetch();
+    }
+
+}
